Added unit tests for observability helper.
I followed a similar approach/pattern as in the TNF Cert Tool project,
creating a slice of unnamed structs for each unit test.

In order to launch UT globally (./...) we need to filter out the ginkgo
TestX funcions, so they must be marked with the  "//go:build !utests"
comment as the first comment in each file. This will also be required on
each new ginkgo test suite added to the project. If we don't want to add
this comment to the test suites, the only option is to rename all those
files to remove the _test suffix.

The UNIT_TEST env var is necessary so the init() from the globalhelper
package (init.go) can avoid the k8s clients init and create an empty
Configuration struct.

Added to github CI.