Avoid setting master field for NADs.
There's no need to use any common interface between workers so it can be
set in network attachment definition. If no "master" field is set, the
NAD uses the default interface on each node, which should be enough for
the purpose of this test cases.

I've left the helper functions that search those "multus interfaces" in
the worker nodes in case they need to be used in the future.