Adjusted operator-install test case.
Certsuite's PR
https://github.com/redhat-best-practices-for-k8s/certsuite/pull/2653
changed the way CSVs are autodiscovered in order for operator pods to be
tested: the operator installation namespace must be explicitly added to
the configured namespaces.

This change deploys the operatorgroup and subscription of the logging
operator in the random test namespace that is set in the certsuite
confi file.