Add log output for debugging when helm installation fails.
Just added the stdout+stderr output of the helm tool installation as the
affiliated test suite seems to fail too often. As this happens at the
SynchronizedBeforeSuite() func, all the test cases are skipped when it
happens.

I've seen this failure in all our OCP versions where we run QE.

Last failing runs in 4.16:
https://github.com/redhat-best-practices-for-k8s/certsuite/actions/runs/12919752505
https://github.com/redhat-best-practices-for-k8s/certsuite/actions/runs/13001158158
https://github.com/redhat-best-practices-for-k8s/certsuite/actions/runs/13220699737